Revert "split up tests"
This reverts commit 7f19149bf1a98e47a31b785d2a292b1cc9ab7095.
